Section 3 corrected and shortened.
Also contains minor edits on other sections. There are two references
for Newman's lemma now, which are two more than necessary and hopefully
enough. Also added the needspace package, which is necessary to keep
diagrams from bleeding into footnotes or page numbers; however it needs
to be used consistently, which it so far is not.