Cosmetic changes on Section 7
Pavel, does \alpha look good for the "ceq"? I don't have any particular
preference, but "ceq" looks really confusing given that there already is
an operator \operatorname{ceq}.
I had to get rid of a figure environment to make leftbar work. When we
submit to a journal, we will get rid of the leftbars, but I'd prefer to
keep them so far (they show where exactly the environments end, and make
them easy to find).