Section 1 written, and Section 5 shortened and corrected.
This does not mean that Section 1 is actually a good introduction, nor
that Section 5 is actually short. But I think the paper is now readable
without looking like an unfinished building or insulting the reader's
intelligence. I think further shortness can be gained by inlining the
diamond lemma, removing the long equations in Proposition 5.8 (d),
shortening examples, etc., but we should see which of these bridges we
want to cross.