Starting to edit Section 7.
I use "an rpp" rather than "a rpp". Both are correct, but the latter
makes me stutter when I read the text and I want to be consistent; if
you have a strong opinion against it, I'll change it the other way
around throughout the text.

Replaced "skew-shapes" by "skew partitions". Might not always be
correct, as there is a minor difference between the two notions, but it
will become clear when I'm reading the appropriate part of the argument.

Also, more minor changes.