Section 4: revert some edits
Still not happy with Case 1: I think the old proof was clearer. It was
something you could read and tick off, rather than having to convince
yourself by drawing pictures in your mind and inventing arguments. On
the picture, it is clear that v is a descent of B, but is this actually
true? It is, because there was a separating line in column v of A which
was higher than the lower border of the v+1-th column; thus the two
columns must overlap. But this is not self-evident.

The footnote has gotten shorter but is still there. See the comment.

Removed the \items before the "Otherwise"s. To me, an \item in algorithm
implies seriality, like a semicolon in a programming language. There are
no semicolons in "if b then e1 else e2 endif".

Great job on Section 4, Pavel! I tried to summarize the algorithm, but
never managed to. This might well simplify Section 6.