edits on Section 7 and others
Added in various arguments that were left tacit in Section 7. Still not
done with it, as I don't understand the last proof.

Removed some command declarations because they were doubled.

Please don't mimic my \operatorname*{blah}\nolimits syntax; it is
nonsense. The * means "put sub/superscripts below the operator", and the
\nolimits means "no, don't put sub/superscripts below the operator". So
the two things cancel out each other. The reason why I have them both so
often is that this is the code Scientific WorkPlace auto-generates.