response edits to Pavel's 6 commits
I think some of the arguments do deserve more detail, so I've reinserted
some of that again. Most of the changes have been preserved. We do talk
of the "ceq statistic" and the "ircont statistic", so the definitions
are necessary, but what do you think about dropping the word "statistic"
everywhere (particularly since it is not a single integer)?

Haven't looked at the structure section; you definitely know better what
to do about it, Pavel.